API/API/requirements.txt
cloudinary==1.40.0
fastapi==0.111.0
gradio==4.32.2
gradio_client==0.17.0
numpy==1.26.4
openai==1.31.0
python-dotenv==1.0.1
Requests==2.32.3
